+++ src/OrleansServiceBus/OrleansServiceBus.csproj
@@ -51,6 +51,7 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Providers\EventDataGeneratorStreamProvider\EventDataGeneratorStreamProvider.cs" />
     <Compile Include="Providers\EventDataGeneratorStreamProvider\EeventHubPartitionGeneratorReceiver.cs" />
+    <Compile Include="Providers\EventDataGeneratorStreamProvider\EventHubPartitionDataGenerator.cs" />
     <Compile Include="Providers\EventDataGeneratorStreamProvider\EventHubUtils.cs" />
     <Compile Include="Providers\EventDataGeneratorStreamProvider\IEventDataGenerator.cs" />
     <Compile Include="Providers\Streams\EventHub\CachePressureMonitors\AggregatedCachePressureMonitor.cs" />
