+++ src/OrleansServiceBus/Providers/EventDataGeneratorStreamProvider/IEventDataGenerator.cs
@@ -14,7 +14,7 @@ namespace Orleans.ServiceBus.Providers
 {
     internal interface IDataGenerator<T>
     {
-        bool TryReadEvents(int maxCount, TimeSpan waitTime, out IEnumerable<T> events);
+        bool TryReadEvents(int maxCount, out IEnumerable<T> events);
     }
 
     internal interface IStreamDataGeneratingController
@@ -28,145 +28,6 @@ namespace Orleans.ServiceBus.Providers
         IntCounter SequenceNumberCounter { set; }
         IStreamIdentity StreamId { get; }
         bool ShouldProduce { set; }
-    }
-
-    /// <summary>
-    /// Generate data for one stream
-    /// </summary>
-    internal class SimpleStreamEventDataGenerator : IStreamDataGenerator<EventData>
-    {
-        public IStreamIdentity StreamId { get; set; }
-        public IntCounter SequenceNumberCounter { set; private get; }
-        private Logger logger;
-        public bool ShouldProduce { private get; set; }
-        private SerializationManager serializationManager;
-        private int drainEventCount;
-        public SimpleStreamEventDataGenerator(IStreamIdentity streamId, EventHubGeneratorStreamProviderSettings settings, 
-            Logger logger, SerializationManager serializationManager)
-        {
-            this.StreamId = streamId;
-            this.logger = logger.GetSubLogger(this.GetType().Name);
-            this.ShouldProduce = true;
-            this.serializationManager = serializationManager;
-            this.drainEventCount = settings.DrainEventCount;
-        }
-
-        public bool TryReadEvents(int eventCount, TimeSpan waitTime, out IEnumerable<EventData> events)
-        {
-            if (!this.ShouldProduce)
-            {
-                events = null;
-                return false;
-            }
-            int count = GetEventGeneratingCount(eventCount);
-            List<EventData> eventDataList = new List<EventData>();
-            while (count-- > 0)
-            {
-                this.SequenceNumberCounter.Increment();
-                if(this.drainEventCount > 0)
-                    this.drainEventCount --;
-                var eventData = EventHubBatchContainer.ToEventData<int>(this.serializationManager, this.StreamId.Guid, this.StreamId.Namespace,
-                    this.GenerateEvent(this.SequenceNumberCounter.Value), RequestContext.Export(this.serializationManager));
-#if NETSTANDARD
-                //set partition key
-                eventData.SetPartitionKey(this.StreamId.Guid.ToString());
-#endif
-                //set offset
-                DateTime now = DateTime.UtcNow;
-                var offSet = this.StreamId.Guid.ToString() + now.ToString();
-                eventData.SetOffset(offSet);
-                //set sequence number
-                eventData.SetSequenceNumber(this.SequenceNumberCounter.Value);
-                //set enqueue time
-                eventData.SetEnqueuedTimeUtc(now);
-                eventDataList.Add(eventData);
-                this.logger.Info($"Generate data of SequemceNumber {SequenceNumberCounter.Value} for stream {this.StreamId.Namespace}-{this.StreamId.Guid}");
-            }
-
-            events = eventDataList;
-            return eventDataList.Count > 0;
-        }
-
-
-        private int GetEventGeneratingCount(int eventCount)
-        {
-            if (this.drainEventCount > 0)
-            {
-                //finish producing drain events first
-                return this.drainEventCount > eventCount ? eventCount : this.drainEventCount;
-            }
-            else
-            {
-                //after producing drain events, producing asked event count
-                return eventCount;
-            }
-        }
-
-        IEnumerable<int> GenerateEvent(int sequenceNumber)
-        {
-            var events = new List<int>();
-            events.Add(sequenceNumber);
-            return events;
-        }
-    }
-
-    /// <summary>
-    /// EHPartitionDataGenerator generate data for a EH partition, which can include data from different streams
-    /// </summary>
-    internal class EventHubPartitionDataGenerator : IDataGenerator<EventData>, IStreamDataGeneratingController
-    {
-        //differnt stream in the same partition should use the same sequenceNumberCounter
-        private IntCounter sequenceNumberCounter = new IntCounter();
-        private Logger logger;
-        private List<IStreamDataGenerator<EventData>> generators;
-        private SerializationManager serializationManager;
-        private EventHubGeneratorStreamProviderSettings settings;
-        private Random seed;
-        public EventHubPartitionDataGenerator(Logger logger, SerializationManager serializationManager, EventHubGeneratorStreamProviderSettings settings)
-        {
-            this.logger = logger.GetSubLogger(this.GetType().Name);
-            this.generators = new List<IStreamDataGenerator<EventData>>();
-            this.serializationManager = serializationManager;
-            this.settings = settings;
-            this.seed = new Random();
-        }
-        public void AddDataGeneratorForStream(IStreamIdentity streamId)
-        {
-            var generator = (IStreamDataGenerator<EventData>)Activator.CreateInstance(settings.StreamDataGeneratorType,
-                streamId, settings, this.logger, this.serializationManager);
-            generator.SequenceNumberCounter = sequenceNumberCounter;
-            this.logger.Info($"Data generator set up on stream {streamId.Namespace}-{streamId.Guid.ToString()}.");
-            this.generators.Add(generator);
-        }
-        public void StopProducingOnStream(IStreamIdentity streamId)
-        {
-            this.generators.ForEach(generator => {
-                if (generator.StreamId.Equals(streamId))
-                {
-                    generator.ShouldProduce = false;
-                    this.logger.Info($"Stop producing data on stream {streamId.Namespace}-{streamId.Guid.ToString()}.");
-                }
-            });
-        }
-        public bool TryReadEvents(int maxCount, TimeSpan waitTime, out IEnumerable<EventData> events)
-        {
-            var eventDataList = new List<EventData>();
-            int remainingMaxCount = maxCount;
-            this.generators.ForEach(generator => {
-                IEnumerable<EventData> eventData;
-                int count = this.seed.Next(remainingMaxCount + 1);
-                remainingMaxCount -= count;
-                if (generator.TryReadEvents(count, waitTime, out eventData))
-                {
-                    foreach (var data in eventData)
-                    {
-                        eventDataList.Add(data);
-                    }
-                }
-            });
-            events = eventDataList.AsEnumerable();
-            return eventDataList.Count > 0;
-        }
     }
 
     internal class IntCounter
